WIP for Python 3 on darwin-* and mingw
For darwin, it's throwing errors in system headers while compiling
Modules/posixmodule.c.
mingw requires at least a toolchain update to provide `wcsnlen_s'
and with a bit of patching, it's now compiling all source files and
fails when linking. The reason seems to be that there are some files
in python3-3.7.4/PC that provide symbols actually needed by Modules/.
However, they are only referenced by the project files in PCbuild/,
but not compiled by the provided Makefiles :-(